add filtered view support for topology table
