+++ deepfence_frontend/apps/dashboard/src/features/topology/pages/Graph.tsx
@@ -1,11 +1,11 @@
 import { useSearchParams } from 'react-router-dom';
 
-import { Graph as GraphComponent } from '@/features/topology/data-components/Graph';
+import { TopologyGraph } from '@/features/topology/components/TopologyGraph';
 
 const Graph = () => {
   const [searchParams] = useSearchParams();
   const type = searchParams.get('type') ?? 'cloud';
-  return <GraphComponent key={type} />;
+  return <TopologyGraph key={type} />;
 };
 
 export const module = {
