+++ deepfence_frontend/apps/dashboard/src/features/topology/pages/Topology.tsx
@@ -6,7 +6,7 @@ function Topology() {
   return (
     <div className="h-full flex flex-col">
       <TopologyHeader />
-      <div className="m-2 flex-1">
+      <div className="p-2 flex-1">
         <Outlet />
       </div>
     </div>
