+++ deepfence_frontend/apps/dashboard/src/routes/private.tsx
@@ -65,9 +65,9 @@ import { module as settings } from '@/features/settings/pages/Settings';
 import { module as userManagement } from '@/features/settings/pages/UserManagement';
 import { module as threatGraphDetailModal } from '@/features/threat-graph/data-components/DetailsModal';
 import { module as threatGraph } from '@/features/threat-graph/pages/ThreatGraph';
-import { module as topologyGraphAction } from '@/features/topology/data-components/Graph';
 import { module as nodeDetailsContainer } from '@/features/topology/data-components/node-details/Container';
 import { module as nodeDetailsHost } from '@/features/topology/data-components/node-details/Host';
+import { module as topologyAction } from '@/features/topology/data-components/topologyAction';
 import { module as topologyGraph } from '@/features/topology/pages/Graph';
 import { module as topologyTable } from '@/features/topology/pages/Table';
 import { module as topology } from '@/features/topology/pages/Topology';
@@ -480,8 +480,8 @@ export const privateRoutes: CustomRouteObject[] = [
         ...threatGraphDetailModal,
       },
       {
-        path: 'topology/graph',
-        ...topologyGraphAction,
+        path: 'topology',
+        ...topologyAction,
       },
     ],
   },
